refactor: split handler.go into service.go + handler.go
- service.go: SessionProvider implementation (ListSessions, GetSession,
  UpdateMeta, HandleTerminalWS) — business logic
- handler.go: HTTP handlers only (handleList, handleGet, handleCreate,
  handleDelete, handleSwitchMode, handleTerminalWS) — routing wrappers

Both files stay in package session. No behavior changes.
